+++ packages/demo/CHANGELOG.md
@@ -1,5 +1,13 @@
 # demo
 
+## 0.1.2
+
+### Patch Changes
+
+- Improving tab navigation, updating documentation, removing bugs with textarea and changing language input type
+- Updated dependencies
+  - @lmarcel/highlight@2.1.0
+
 ## 0.1.1
 
 ### Patch Changes